refactor: remove lib.rs to enable dead code detection in binary crate
- Remove [lib] section from Cargo.toml to make cribo a true binary-only crate
- Move unit tests from separate test files into their respective module files:
  - test_pythonpath_support.rs tests moved to resolver.rs
  - test_python_version_config.rs tests moved to config.rs
  - Delete test_virtualenv_support.rs (can be restored later if needed)
- Update test_bundling_snapshots.rs to use binary execution instead of library imports:
  - Add run_cribo() helper to execute cribo binary
  - Update error handling for binary stderr output
  - Accept 2 snapshot changes for new error format with "Error: " prefix
- Create conditional lib.rs that only exists with 'bench' feature flag:
  - Add [features] bench = [] in Cargo.toml
  - Configure lib with required-features = ["bench"]
  - Update benchmarks to require bench feature
  - Keeps benchmarks working while preserving dead code detection
- Fix missing anyhow::Result import in resolver tests module
- Simplify ecosystem benchmark with TODO placeholder

This change enables proper dead code detection for the binary-only crate,
helping identify unused code that can be removed. The conditional library
approach allows benchmarks to continue working without affecting normal builds.
