+++ crates/cribo/src/lib.rs
@@ -1,3 +1,9 @@
+// This file only exists when the 'bench' feature is enabled
+// It's used exclusively for benchmarking and does not affect dead code detection
+// in normal builds
+
+#![cfg(feature = "bench")]
+
 pub mod ast_indexer;
 pub mod code_generator;
 pub mod combine;
@@ -5,19 +11,14 @@ pub mod config;
 pub mod cribo_graph;
 pub mod dirs;
 pub mod graph_builder;
-pub mod import_alias_tracker;
-pub mod import_rewriter;
+pub mod module_registry;
+pub mod module_transformer;
 pub mod orchestrator;
+pub mod python_compatibility;
 pub mod resolver;
 pub mod semantic_analysis;
-pub mod semantic_bundler;
-pub mod side_effects;
-pub mod stdlib_normalization;
 pub mod transformation_context;
 pub mod tree_shaking;
-pub mod types;
+pub mod unused_imports;
 pub mod util;
 pub mod visitors;
-
-pub use config::Config;
-pub use orchestrator::BundleOrchestrator;
