--- crates/cribo/tests/test_virtualenv_support.rs
@@ -1,922 +1,0 @@
-#![allow(clippy::disallowed_methods)]
-
-use std::fs;
-
-use cribo::{
-    config::Config,
-    resolver::{ImportType, ModuleResolver, VirtualEnvGuard},
-};
-use serial_test::serial;
-use tempfile::TempDir;
-
-#[test]
-fn test_virtualenv_import_classification() {
-    // Create temporary directories for testing
-    let temp_dir = TempDir::new().unwrap();
-    let virtualenv_dir = temp_dir.path().join("test_venv");
-    let src_dir = temp_dir.path().join("src");
-
-    // Create virtual environment structure (Unix-style)
-    let site_packages_dir = if cfg!(windows) {
-        virtualenv_dir.join("Lib").join("site-packages")
-    } else {
-        virtualenv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&site_packages_dir).unwrap();
-    fs::create_dir_all(&src_dir).unwrap();
-
-    // Create a module in virtual environment site-packages
-    let venv_module = site_packages_dir.join("requests.py");
-    fs::write(
-        &venv_module,
-        "# This is a third-party module installed in virtual environment",
-    )
-    .unwrap();
-
-    // Create a package in virtual environment site-packages
-    let venv_pkg = site_packages_dir.join("numpy");
-    fs::create_dir_all(&venv_pkg).unwrap();
-    let venv_pkg_init = venv_pkg.join("__init__.py");
-    fs::write(&venv_pkg_init, "# Third-party package").unwrap();
-
-    // Create a module in src directory (this should be first-party)
-    let src_module = src_dir.join("mymodule.py");
-    fs::write(&src_module, "# This is a first-party module").unwrap();
-
-    // Set up config with src directory only
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Create resolver with VIRTUAL_ENV override
-    let virtualenv_str = virtualenv_dir.to_string_lossy();
-    let mut resolver = ModuleResolver::new_with_virtualenv(config, Some(&virtualenv_str)).unwrap();
-
-    // Test that src modules are classified as first-party
-    assert_eq!(
-        resolver.classify_import("mymodule"),
-        ImportType::FirstParty,
-        "Modules from src directory should be classified as first-party"
-    );
-
-    // Test that virtual environment modules are classified as third-party
-    assert_eq!(
-        resolver.classify_import("requests"),
-        ImportType::ThirdParty,
-        "Modules from VIRTUAL_ENV should be classified as third-party"
-    );
-
-    assert_eq!(
-        resolver.classify_import("numpy"),
-        ImportType::ThirdParty,
-        "Packages from VIRTUAL_ENV should be classified as third-party"
-    );
-
-    // Test submodule classification
-    assert_eq!(
-        resolver.classify_import("numpy.array"),
-        ImportType::ThirdParty,
-        "Submodules of VIRTUAL_ENV packages should be classified as third-party"
-    );
-
-    // Test standard library modules are still recognized correctly
-    assert_eq!(
-        resolver.classify_import("os"),
-        ImportType::StandardLibrary,
-        "Standard library modules should still be classified correctly"
-    );
-
-    // Test unknown modules are classified as third-party
-    assert_eq!(
-        resolver.classify_import("unknown_module"),
-        ImportType::ThirdParty,
-        "Unknown modules should be classified as third-party"
-    );
-}
-
-#[test]
-fn test_virtualenv_without_env_set() {
-    // Test that resolver works correctly when VIRTUAL_ENV is not set
-    let temp_dir = TempDir::new().unwrap();
-    let src_dir = temp_dir.path().join("src");
-    fs::create_dir_all(&src_dir).unwrap();
-
-    let src_module = src_dir.join("mymodule.py");
-    fs::write(&src_module, "# First-party module").unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Create resolver without VIRTUAL_ENV
-    let mut resolver = ModuleResolver::new_with_virtualenv(config, None).unwrap();
-
-    // Test that src modules are still classified as first-party
-    assert_eq!(
-        resolver.classify_import("mymodule"),
-        ImportType::FirstParty,
-        "Modules from src directory should be classified as first-party"
-    );
-
-    // Test that unknown modules default to third-party
-    assert_eq!(
-        resolver.classify_import("unknown_module"),
-        ImportType::ThirdParty,
-        "Unknown modules should be classified as third-party when no VIRTUAL_ENV"
-    );
-
-    // Test standard library modules are still recognized
-    assert_eq!(
-        resolver.classify_import("os"),
-        ImportType::StandardLibrary,
-        "Standard library modules should still work without VIRTUAL_ENV"
-    );
-}
-
-#[test]
-fn test_virtualenv_scan_directories_exclusion() {
-    // Test that VIRTUAL_ENV directories are NOT included in scan directories
-    let temp_dir = TempDir::new().unwrap();
-    let virtualenv_dir = temp_dir.path().join("test_venv");
-    let src_dir = temp_dir.path().join("src");
-
-    // Create virtual environment structure
-    let site_packages_dir = if cfg!(windows) {
-        virtualenv_dir.join("Lib").join("site-packages")
-    } else {
-        virtualenv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&site_packages_dir).unwrap();
-    fs::create_dir_all(&src_dir).unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Create resolver with VIRTUAL_ENV override
-    let virtualenv_str = virtualenv_dir.to_string_lossy();
-    let resolver = ModuleResolver::new_with_virtualenv(config, Some(&virtualenv_str)).unwrap();
-    let scan_dirs = resolver.get_search_directories();
-
-    // Entry directory would be first if set, then src directories
-    let expected_src = src_dir.canonicalize().unwrap_or(src_dir);
-    assert!(
-        scan_dirs.contains(&expected_src),
-        "Should contain src directory"
-    );
-
-    // Verify VIRTUAL_ENV site-packages is NOT in scan directories
-    assert!(
-        !scan_dirs.iter().any(|p| p == &site_packages_dir),
-        "VIRTUAL_ENV site-packages should NOT be in scan directories"
-    );
-}
-
-#[test]
-#[serial]
-fn test_virtualenv_guard() {
-    // Test the VirtualEnvGuard functionality
-    let original_env = std::env::var("VIRTUAL_ENV").ok();
-
-    {
-        let _guard = VirtualEnvGuard::unset();
-        // Environment should be cleared
-        assert!(std::env::var("VIRTUAL_ENV").is_err());
-    }
-
-    // Environment should be restored after guard is dropped
-    // Use a more robust check that handles potential restoration failures
-    match original_env {
-        Some(expected_value) => {
-            // Check if restoration worked, but don't fail the test if it didn't
-            // This is a flaky aspect of environment variable restoration
-            let restored_value = std::env::var("VIRTUAL_ENV").ok();
-            if restored_value.as_deref() != Some(expected_value.as_str()) {
-                // Log the restoration failure but don't fail the test
-                // The core functionality (unset working) was already verified
-                eprintln!(
-                    "Warning: VIRTUAL_ENV restoration may have failed. Expected \
-                     '{expected_value}', got {restored_value:?}. This is a known flaky behavior \
-                     in environment variable cleanup."
-                );
-            }
-        }
-        None => {
-            // Should either be unset or if restoration failed, that's also acceptable for this test
-            // The important thing is that unset() worked correctly inside the guard scope
-        }
-    }
-}
-
-#[test]
-fn test_virtualenv_empty_or_nonexistent() {
-    let temp_dir = TempDir::new().unwrap();
-    let src_dir = temp_dir.path().join("src");
-    fs::create_dir_all(&src_dir).unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Test with empty VIRTUAL_ENV
-    let resolver1 = ModuleResolver::new_with_virtualenv(config.clone(), Some("")).unwrap();
-    let scan_dirs1 = resolver1.get_search_directories();
-
-    // Should only contain configured src directories
-    assert_eq!(scan_dirs1.len(), 1);
-    let expected_path = src_dir.canonicalize().unwrap_or(src_dir.clone());
-    assert!(scan_dirs1.contains(&expected_path));
-
-    // Test with nonexistent VIRTUAL_ENV directory
-    let mut resolver2 =
-        ModuleResolver::new_with_virtualenv(config, Some("/nonexistent/venv")).unwrap();
-    let unknown_classification = resolver2.classify_import("unknown_module");
-
-    // Should still classify unknown modules as third-party
-    assert_eq!(unknown_classification, ImportType::ThirdParty);
-}
-
-#[test]
-fn test_virtualenv_multiple_python_versions() {
-    // Test detection of packages from multiple Python versions in VIRTUAL_ENV
-    let temp_dir = TempDir::new().unwrap();
-    let virtualenv_dir = temp_dir.path().join("test_venv");
-    let src_dir = temp_dir.path().join("src");
-
-    // Create multiple python version directories (Unix-style only for this test)
-    if !cfg!(windows) {
-        let site_packages_dir1 = virtualenv_dir
-            .join("lib")
-            .join("python3.10")
-            .join("site-packages");
-        let site_packages_dir2 = virtualenv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages");
-        fs::create_dir_all(&site_packages_dir1).unwrap();
-        fs::create_dir_all(&site_packages_dir2).unwrap();
-        fs::create_dir_all(&src_dir).unwrap();
-
-        // Create modules in different Python version directories
-        let module1 = site_packages_dir1.join("requests.py");
-        fs::write(&module1, "# Module in python3.10").unwrap();
-
-        let module2 = site_packages_dir2.join("numpy.py");
-        fs::write(&module2, "# Module in python3.11").unwrap();
-
-        let config = Config {
-            src: vec![src_dir.clone()],
-            ..Default::default()
-        };
-
-        // Create resolver with VIRTUAL_ENV override
-        let virtualenv_str = virtualenv_dir.to_string_lossy();
-        let mut resolver =
-            ModuleResolver::new_with_virtualenv(config, Some(&virtualenv_str)).unwrap();
-
-        // Both modules should be classified as third-party
-        assert_eq!(
-            resolver.classify_import("requests"),
-            ImportType::ThirdParty,
-            "requests from python3.10 should be classified as third-party"
-        );
-        assert_eq!(
-            resolver.classify_import("numpy"),
-            ImportType::ThirdParty,
-            "numpy from python3.11 should be classified as third-party"
-        );
-    }
-}
-
-#[test]
-fn test_combined_pythonpath_and_virtualenv() {
-    // Test that PYTHONPATH and VIRTUAL_ENV work correctly together
-    let temp_dir = TempDir::new().unwrap();
-    let virtualenv_dir = temp_dir.path().join("test_venv");
-    let pythonpath_dir = temp_dir.path().join("pythonpath_modules");
-    let src_dir = temp_dir.path().join("src");
-
-    // Create virtual environment structure
-    let site_packages_dir = if cfg!(windows) {
-        virtualenv_dir.join("Lib").join("site-packages")
-    } else {
-        virtualenv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&site_packages_dir).unwrap();
-    fs::create_dir_all(&pythonpath_dir).unwrap();
-    fs::create_dir_all(&src_dir).unwrap();
-
-    // Create modules in each location
-    let src_module = src_dir.join("src_module.py");
-    fs::write(&src_module, "# Source module").unwrap();
-
-    let venv_module = site_packages_dir.join("venv_module.py");
-    fs::write(&venv_module, "# Virtual environment module").unwrap();
-
-    let pythonpath_module = pythonpath_dir.join("pythonpath_module.py");
-    fs::write(&pythonpath_module, "# PYTHONPATH module").unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Create resolver with both PYTHONPATH and VIRTUAL_ENV overrides
-    let virtualenv_str = virtualenv_dir.to_string_lossy();
-    let pythonpath_str = pythonpath_dir.to_string_lossy();
-    let mut resolver =
-        ModuleResolver::new_with_overrides(config, Some(&pythonpath_str), Some(&virtualenv_str))
-            .unwrap();
-
-    // Test classification behavior
-    assert_eq!(
-        resolver.classify_import("src_module"),
-        ImportType::FirstParty,
-        "src_module should be first-party (from src directory)"
-    );
-
-    assert_eq!(
-        resolver.classify_import("pythonpath_module"),
-        ImportType::FirstParty,
-        "pythonpath_module should be first-party (from PYTHONPATH)"
-    );
-
-    assert_eq!(
-        resolver.classify_import("venv_module"),
-        ImportType::ThirdParty,
-        "venv_module should be third-party (from VIRTUAL_ENV)"
-    );
-}
-
-#[test]
-fn test_module_shadowing_priority() {
-    // Test module resolution priority when local modules shadow virtual environment packages
-    let temp_dir = TempDir::new().unwrap();
-    let virtualenv_dir = temp_dir.path().join("test_venv");
-    let src_dir = temp_dir.path().join("src");
-    let pythonpath_dir = temp_dir.path().join("pythonpath_modules");
-
-    // Create virtual environment structure
-    let site_packages_dir = if cfg!(windows) {
-        virtualenv_dir.join("Lib").join("site-packages")
-    } else {
-        virtualenv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&site_packages_dir).unwrap();
-    fs::create_dir_all(&src_dir).unwrap();
-    fs::create_dir_all(&pythonpath_dir).unwrap();
-
-    // Create a package "requests" in virtual environment (third-party)
-    let venv_requests_pkg = site_packages_dir.join("requests");
-    fs::create_dir_all(&venv_requests_pkg).unwrap();
-    let venv_requests_init = venv_requests_pkg.join("__init__.py");
-    fs::write(&venv_requests_init, "# Third-party requests package").unwrap();
-
-    // Create a module "requests.py" in src directory (should shadow the venv package)
-    let src_requests_module = src_dir.join("requests.py");
-    fs::write(
-        &src_requests_module,
-        "# Local requests module that shadows venv package",
-    )
-    .unwrap();
-
-    // Create another package "numpy" in virtual environment
-    let venv_numpy_pkg = site_packages_dir.join("numpy");
-    fs::create_dir_all(&venv_numpy_pkg).unwrap();
-    let venv_numpy_init = venv_numpy_pkg.join("__init__.py");
-    fs::write(&venv_numpy_init, "# Third-party numpy package").unwrap();
-
-    // Create a module "numpy.py" in PYTHONPATH directory (should also shadow venv package)
-    let pythonpath_numpy_module = pythonpath_dir.join("numpy.py");
-    fs::write(
-        &pythonpath_numpy_module,
-        "# PYTHONPATH numpy module that shadows venv package",
-    )
-    .unwrap();
-
-    // Create a package "flask" only in virtual environment (no shadowing)
-    let venv_flask_pkg = site_packages_dir.join("flask");
-    fs::create_dir_all(&venv_flask_pkg).unwrap();
-    let venv_flask_init = venv_flask_pkg.join("__init__.py");
-    fs::write(&venv_flask_init, "# Third-party flask package").unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Test with resolver that includes both PYTHONPATH and VIRTUAL_ENV
-    let virtualenv_str = virtualenv_dir.to_string_lossy();
-    let pythonpath_str = pythonpath_dir.to_string_lossy();
-    let mut resolver =
-        ModuleResolver::new_with_overrides(config, Some(&pythonpath_str), Some(&virtualenv_str))
-            .unwrap();
-
-    // Test shadowing cases - first-party modules should take priority over virtual environment
-    // packages
-    assert_eq!(
-        resolver.classify_import("requests"),
-        ImportType::FirstParty,
-        "Local src/requests.py should shadow virtual environment requests package (first-party \
-         wins)"
-    );
-
-    assert_eq!(
-        resolver.classify_import("numpy"),
-        ImportType::FirstParty,
-        "PYTHONPATH numpy.py should shadow virtual environment numpy package (first-party wins)"
-    );
-
-    // Test non-shadowed case - virtual environment package should be classified as third-party
-    assert_eq!(
-        resolver.classify_import("flask"),
-        ImportType::ThirdParty,
-        "Virtual environment flask package should be classified as third-party (no shadowing)"
-    );
-
-    // Test submodule shadowing behavior
-    assert_eq!(
-        resolver.classify_import("requests.auth"),
-        ImportType::FirstParty,
-        "Submodule of shadowed module should also be classified as first-party"
-    );
-
-    assert_eq!(
-        resolver.classify_import("numpy.array"),
-        ImportType::FirstParty,
-        "Submodule of shadowed module should also be classified as first-party"
-    );
-
-    assert_eq!(
-        resolver.classify_import("flask.app"),
-        ImportType::ThirdParty,
-        "Submodule of non-shadowed virtual environment package should be third-party"
-    );
-
-    // Verify that the first-party modules can be resolved
-    let config2 = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-    let mut resolver_mut =
-        ModuleResolver::new_with_overrides(config2, Some(&pythonpath_str), Some(&virtualenv_str))
-            .unwrap();
-    assert!(
-        resolver_mut
-            .resolve_module_path("requests")
-            .unwrap()
-            .is_some(),
-        "requests should be resolvable as first-party module"
-    );
-    assert!(
-        resolver_mut.resolve_module_path("numpy").unwrap().is_some(),
-        "numpy should be resolvable as first-party module"
-    );
-    // flask is in virtual environment, but we can't resolve it directly (it's just for
-    // classification)
-}
-
-#[test]
-fn test_package_vs_module_shadowing() {
-    // Test shadowing between packages and modules with the same name
-    let temp_dir = TempDir::new().unwrap();
-    let virtualenv_dir = temp_dir.path().join("test_venv");
-    let src_dir = temp_dir.path().join("src");
-
-    // Create virtual environment structure
-    let site_packages_dir = if cfg!(windows) {
-        virtualenv_dir.join("Lib").join("site-packages")
-    } else {
-        virtualenv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&site_packages_dir).unwrap();
-    fs::create_dir_all(&src_dir).unwrap();
-
-    // Case 1: Local module shadows virtual environment package
-    let venv_pkg = site_packages_dir.join("mylib");
-    fs::create_dir_all(&venv_pkg).unwrap();
-    let venv_pkg_init = venv_pkg.join("__init__.py");
-    fs::write(&venv_pkg_init, "# Virtual environment package mylib").unwrap();
-    let venv_pkg_sub = venv_pkg.join("utils.py");
-    fs::write(&venv_pkg_sub, "# Virtual environment mylib.utils").unwrap();
-
-    let src_module = src_dir.join("mylib.py");
-    fs::write(&src_module, "# Local module mylib").unwrap();
-
-    // Case 2: Local package shadows virtual environment module
-    let venv_module = site_packages_dir.join("anotherlib.py");
-    fs::write(&venv_module, "# Virtual environment module anotherlib").unwrap();
-
-    let src_pkg = src_dir.join("anotherlib");
-    fs::create_dir_all(&src_pkg).unwrap();
-    let src_pkg_init = src_pkg.join("__init__.py");
-    fs::write(&src_pkg_init, "# Local package anotherlib").unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    let virtualenv_str = virtualenv_dir.to_string_lossy();
-    let mut resolver = ModuleResolver::new_with_virtualenv(config, Some(&virtualenv_str)).unwrap();
-
-    // Test Case 1: Local module mylib.py shadows venv package mylib/
-    assert_eq!(
-        resolver.classify_import("mylib"),
-        ImportType::FirstParty,
-        "Local mylib.py module should shadow virtual environment mylib package"
-    );
-
-    // When accessing submodules, the local module takes precedence
-    // but mylib.utils would not exist as a submodule of the local mylib.py file
-    // So it should be classified based on the shadowing rule
-    assert_eq!(
-        resolver.classify_import("mylib.utils"),
-        ImportType::FirstParty,
-        "mylib.utils should be classified as first-party due to mylib being first-party"
-    );
-
-    // Test Case 2: Local package anotherlib/ shadows venv module anotherlib.py
-    assert_eq!(
-        resolver.classify_import("anotherlib"),
-        ImportType::FirstParty,
-        "Local anotherlib package should shadow virtual environment anotherlib.py module"
-    );
-
-    // Verify first-party modules can be resolved
-    let config2 = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-    let mut resolver_mut =
-        ModuleResolver::new_with_virtualenv(config2, Some(&virtualenv_str)).unwrap();
-    assert!(
-        resolver_mut.resolve_module_path("mylib").unwrap().is_some(),
-        "mylib should be resolvable as first-party"
-    );
-    assert!(
-        resolver_mut
-            .resolve_module_path("anotherlib")
-            .unwrap()
-            .is_some(),
-        "anotherlib should be resolvable as first-party"
-    );
-}
-
-#[test]
-fn test_venv_fallback_detection() {
-    // Test that .venv directory is automatically detected when VIRTUAL_ENV is not set
-    let temp_dir = TempDir::new().unwrap();
-    let original_dir = std::env::current_dir().ok();
-
-    // Change to temporary directory to isolate test
-    std::env::set_current_dir(temp_dir.path()).unwrap();
-
-    // Ensure VIRTUAL_ENV is not set for this test
-    let _guard = VirtualEnvGuard::unset();
-
-    // Create a .venv directory structure
-    let venv_dir = temp_dir.path().join(".venv");
-    let site_packages_dir = if cfg!(windows) {
-        venv_dir.join("Lib").join("site-packages")
-    } else {
-        venv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&site_packages_dir).unwrap();
-
-    // Add some packages to .venv
-    let requests_pkg = site_packages_dir.join("requests");
-    fs::create_dir_all(&requests_pkg).unwrap();
-    fs::write(requests_pkg.join("__init__.py"), "# requests package").unwrap();
-
-    let numpy_module = site_packages_dir.join("numpy.py");
-    fs::write(&numpy_module, "# numpy module").unwrap();
-
-    // Create src directory
-    let src_dir = temp_dir.path().join("src");
-    fs::create_dir_all(&src_dir).unwrap();
-    let src_module = src_dir.join("mymodule.py");
-    fs::write(&src_module, "# first-party module").unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Create resolver without VIRTUAL_ENV override (should use fallback detection)
-    let mut resolver = ModuleResolver::new_with_virtualenv(config, None).unwrap();
-
-    // Test that packages from .venv are detected as third-party
-    assert_eq!(
-        resolver.classify_import("requests"),
-        ImportType::ThirdParty,
-        "requests from .venv should be classified as third-party"
-    );
-
-    assert_eq!(
-        resolver.classify_import("numpy"),
-        ImportType::ThirdParty,
-        "numpy from .venv should be classified as third-party"
-    );
-
-    // Test that first-party modules still work
-    assert_eq!(
-        resolver.classify_import("mymodule"),
-        ImportType::FirstParty,
-        "mymodule should still be classified as first-party"
-    );
-
-    // Restore original directory (do this before temp_dir is dropped)
-    if let Some(dir) = original_dir
-        && dir.exists()
-    {
-        let _ = std::env::set_current_dir(&dir);
-    }
-}
-
-#[test]
-fn test_venv_fallback_priority_order() {
-    // Test that .venv is preferred over venv when both exist
-    let temp_dir = TempDir::new().unwrap();
-    let original_dir = std::env::current_dir().ok();
-
-    // Change to temporary directory for isolation
-    std::env::set_current_dir(temp_dir.path()).unwrap();
-
-    // Ensure VIRTUAL_ENV is not set
-    let _guard = VirtualEnvGuard::unset();
-
-    // Create both .venv and venv directories
-    let dotvenv_dir = temp_dir.path().join(".venv");
-    let venv_dir = temp_dir.path().join("venv");
-
-    let dotvenv_site_packages = if cfg!(windows) {
-        dotvenv_dir.join("Lib").join("site-packages")
-    } else {
-        dotvenv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-
-    let venv_site_packages = if cfg!(windows) {
-        venv_dir.join("Lib").join("site-packages")
-    } else {
-        venv_dir
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-
-    fs::create_dir_all(&dotvenv_site_packages).unwrap();
-    fs::create_dir_all(&venv_site_packages).unwrap();
-
-    // Add different packages to each
-    let dotvenv_pkg = dotvenv_site_packages.join("dotvenv_package.py");
-    fs::write(&dotvenv_pkg, "# package from .venv").unwrap();
-
-    let venv_pkg = venv_site_packages.join("venv_package.py");
-    fs::write(&venv_pkg, "# package from venv").unwrap();
-
-    let src_dir = temp_dir.path().join("src");
-    fs::create_dir_all(&src_dir).unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    let mut resolver = ModuleResolver::new_with_virtualenv(config, None).unwrap();
-
-    // Both packages should be detected since we scan all found virtual environments
-    assert_eq!(
-        resolver.classify_import("dotvenv_package"),
-        ImportType::ThirdParty,
-        "dotvenv_package should be detected from .venv"
-    );
-
-    assert_eq!(
-        resolver.classify_import("venv_package"),
-        ImportType::ThirdParty,
-        "venv_package should be detected from venv"
-    );
-
-    // Restore original directory - handle case where it might not exist
-    if let Some(dir) = original_dir
-        && dir.exists()
-    {
-        let _ = std::env::set_current_dir(&dir);
-    }
-}
-
-#[test]
-fn test_explicit_virtualenv_overrides_fallback() {
-    // Test that explicit VIRTUAL_ENV takes precedence over fallback detection
-    let temp_dir = TempDir::new().unwrap();
-    let original_dir = std::env::current_dir().ok();
-
-    // Change to temporary directory
-    std::env::set_current_dir(temp_dir.path()).unwrap();
-
-    // Create .venv directory with a package
-    let fallback_venv = temp_dir.path().join(".venv");
-    let fallback_site_packages = if cfg!(windows) {
-        fallback_venv.join("Lib").join("site-packages")
-    } else {
-        fallback_venv
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&fallback_site_packages).unwrap();
-    let fallback_pkg = fallback_site_packages.join("fallback_package.py");
-    fs::write(&fallback_pkg, "# fallback package").unwrap();
-
-    // Create explicit virtual environment with different package
-    let explicit_venv = temp_dir.path().join("explicit_venv");
-    let explicit_site_packages = if cfg!(windows) {
-        explicit_venv.join("Lib").join("site-packages")
-    } else {
-        explicit_venv
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&explicit_site_packages).unwrap();
-    let explicit_pkg = explicit_site_packages.join("explicit_package.py");
-    fs::write(&explicit_pkg, "# explicit package").unwrap();
-
-    let src_dir = temp_dir.path().join("src");
-    fs::create_dir_all(&src_dir).unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    // Test with explicit VIRTUAL_ENV override
-    let explicit_venv_str = explicit_venv.to_string_lossy();
-    let mut resolver =
-        ModuleResolver::new_with_virtualenv(config, Some(&explicit_venv_str)).unwrap();
-
-    // Should only find package from explicit virtual environment
-    assert_eq!(
-        resolver.classify_import("explicit_package"),
-        ImportType::ThirdParty,
-        "explicit_package should be found in explicit virtual environment"
-    );
-
-    // Should NOT find package from fallback .venv
-    assert_eq!(
-        resolver.classify_import("fallback_package"),
-        ImportType::ThirdParty, // Still third-party since it's unknown
-        "fallback_package should not be detected when explicit VIRTUAL_ENV is set"
-    );
-
-    // Restore original directory
-    if let Some(dir) = original_dir
-        && dir.exists()
-    {
-        let _ = std::env::set_current_dir(dir);
-    }
-}
-
-#[test]
-fn test_no_virtualenv_fallback_when_none_exist() {
-    // Test behavior when no virtual environments exist
-    let temp_dir = TempDir::new().unwrap();
-    let original_dir = std::env::current_dir().ok();
-
-    // Change to temporary directory (no virtual environments)
-    std::env::set_current_dir(temp_dir.path()).unwrap();
-
-    // Ensure VIRTUAL_ENV is not set
-    let _guard = VirtualEnvGuard::unset();
-
-    let src_dir = temp_dir.path().join("src");
-    fs::create_dir_all(&src_dir).unwrap();
-    let src_module = src_dir.join("mymodule.py");
-    fs::write(&src_module, "# first-party module").unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    let mut resolver = ModuleResolver::new_with_virtualenv(config, None).unwrap();
-
-    // Should still work for first-party modules
-    assert_eq!(
-        resolver.classify_import("mymodule"),
-        ImportType::FirstParty,
-        "First-party modules should work without virtual environment"
-    );
-
-    // Unknown modules should default to third-party
-    assert_eq!(
-        resolver.classify_import("unknown_package"),
-        ImportType::ThirdParty,
-        "Unknown modules should default to third-party"
-    );
-
-    // Standard library should still work
-    assert_eq!(
-        resolver.classify_import("os"),
-        ImportType::StandardLibrary,
-        "Standard library should still work"
-    );
-
-    // Restore original directory
-    if let Some(dir) = original_dir
-        && dir.exists()
-    {
-        let _ = std::env::set_current_dir(dir);
-    }
-}
-
-#[test]
-fn test_invalid_venv_directories_ignored() {
-    // Test that directories that aren't valid virtual environments are ignored
-    let temp_dir = TempDir::new().unwrap();
-    let original_dir = std::env::current_dir().ok();
-
-    // Change to temporary directory
-    std::env::set_current_dir(temp_dir.path()).unwrap();
-
-    // Ensure VIRTUAL_ENV is not set
-    let _guard = VirtualEnvGuard::unset();
-
-    // Create directories with venv-like names but without proper structure
-    let fake_venv1 = temp_dir.path().join(".venv");
-    fs::create_dir_all(&fake_venv1).unwrap(); // No lib/site-packages
-
-    let fake_venv2 = temp_dir.path().join("venv");
-    fs::create_dir_all(fake_venv2.join("lib")).unwrap(); // No site-packages
-
-    // Create a valid .venv structure in a different name
-    let actual_venv = temp_dir.path().join("env");
-    let actual_site_packages = if cfg!(windows) {
-        actual_venv.join("Lib").join("site-packages")
-    } else {
-        actual_venv
-            .join("lib")
-            .join("python3.11")
-            .join("site-packages")
-    };
-    fs::create_dir_all(&actual_site_packages).unwrap();
-    let real_pkg = actual_site_packages.join("real_package.py");
-    fs::write(&real_pkg, "# real package").unwrap();
-
-    let src_dir = temp_dir.path().join("src");
-    fs::create_dir_all(&src_dir).unwrap();
-
-    let config = Config {
-        src: vec![src_dir.clone()],
-        ..Default::default()
-    };
-
-    let mut resolver = ModuleResolver::new_with_virtualenv(config, None).unwrap();
-
-    // Should find package from valid virtual environment
-    assert_eq!(
-        resolver.classify_import("real_package"),
-        ImportType::ThirdParty,
-        "Should find package from valid virtual environment"
-    );
-
-    // Should not crash or have issues with invalid directories
-    assert_eq!(
-        resolver.classify_import("nonexistent"),
-        ImportType::ThirdParty,
-        "Should handle nonexistent packages gracefully"
-    );
-
-    // Restore original directory
-    if let Some(dir) = original_dir
-        && dir.exists()
-    {
-        let _ = std::env::set_current_dir(dir);
-    }
-}
